fix: add pkcs7 option support for authenticode signing
Fixes: #19

We use a certificate chain off of the DB key also to sign our Secure
Boot binaries - we hit this limitation as we use go-uefi as a library
to sign them.

This patch has been tested and it produces a Secure Boot valid binary
that is accepted by EDK2 (we have tested it with VMs so far).

I also added it to the gosign binary, but we do not use this ourselves
so that code path is untested.